# pkg/api/openapi.yaml
openapi: 3.1.0
info:
  title: OffGrid Local API
  version: 1.0.0
  description: Stable contract shared by the OffGrid service, web UI, desktop app, CLI, and external agents.
servers:
  - url: http://127.0.0.1:11611
security:
  - sessionCookie: []
  - bearerAuth: []
paths:
...
  /v1/agents/tasks:
    get:
      operationId: listAgentTasks
      responses:
        '200':
          description: Durable agent task states
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/AgentTask' } }
        '503': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
    bearerAuth:
      type: http
      scheme: bearer
...
  responses:
    Error:
      description: API error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          oneOf:
            - type: string
            - type: object
              required: [message]
              properties:
                message: { type: string }
                type: { type: string }
                code: { type: string }
...
    ToolApproval:
      type: object
      required: [id, run_id, call_id, actor, tool, arguments, canonical_arguments, expires_at]
      properties:
        id: { type: string }
        run_id: { type: string }
        call_id: { type: string }
        actor: { type: string }
        tool: { type: string }
        arguments: { type: object, additionalProperties: true }
        canonical_arguments: { type: string, description: Exact canonical JSON for review without JavaScript numeric precision loss. }
        capability: { type: object, additionalProperties: true }
        expires_at: { type: string, format: date-time }
...
    AgentStep:
      type: object
      properties:
        id: { type: integer }
        type: { type: string }
        content: { type: string }
        tool_name: { type: string }
        tool_args: { type: string }
        tool_result: { type: string }
        authorization: { type: string, description: Durable record of exact approval or the immutable automatic session policy used for this action. }
        timestamp: { type: string, format: date-time }
...
    AgentTask:
      type: object
      required: [id, prompt, status, created_at]
      properties:
        deletable: { type: boolean, description: Whether this actor may remove this terminal run or interrupted record without recovery state; the server rechecks ownership and execution state at deletion time. }
        id: { type: string }
        parent_id: { type: string }
        prompt: { type: string }
        status: { type: string }
        model: { type: string }
        actor: { type: string }
        pending_approval: { $ref: '#/components/schemas/ToolApproval' }
        result: { type: string }
        error: { type: string }
        error_code: { type: string, description: Stable machine-readable failure code for recovery and automation. }
        steps: { type: array, items: { $ref: '#/components/schemas/AgentStep' } }
        created_at: { type: string, format: date-time }
        started_at: { type: string, format: date-time }
        completed_at: { type: string, format: date-time }
...